% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkPulse.R
\name{checkPulse}
\alias{checkPulse}
\title{checkPulse}
\usage{
checkPulse(cl = NULL)
}
\arguments{
\item{cl}{The cluster object}
}
\value{
A logical vector by list order of the nodes in the cluster object and whether they responded
}
\description{
Nudges each node to see if the connection is alive
}

