% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/healCluster.R
\name{healCluster}
\alias{healCluster}
\title{healCluster}
\usage{
healCluster(cl)
}
\arguments{
\item{cl}{The cluster object}
}
\value{
cl The healed cluster object
}
\description{
Repeatedly checks the status of nodes, disables the first inactive node,
transitions the identity of the last active node to the first inactive node
then disables the identity of the last active node so that it only
exists once in the cluster list.
}

